requirements.txt
langchain
python-dotenv
langchain-google-genai
gradio
django
djangorestframework
django-phonenumber-field[phonenumbers]
google-generativeai
requests
pytesseract
pillow
